Split rita common loop into fast and slow
