+++ rita/src/exit.rs
@@ -45,6 +45,9 @@ mod middleware;
 mod rita_common;
 mod rita_exit;
 
+use rita_common::rita_loop::fast_loop::RitaFastLoop;
+use rita_common::rita_loop::slow_loop::RitaSlowLoop;
+
 use crate::rita_common::dashboard::babel::*;
 use crate::rita_common::dashboard::dao::*;
 use crate::rita_common::dashboard::debts::*;
@@ -275,7 +278,8 @@ fn main() {
     .shutdown_timeout(0)
     .start();
 
-    assert!(rita_common::rita_loop::RitaLoop::from_registry().connected());
+    assert!(RitaFastLoop::from_registry().connected());
+    assert!(RitaSlowLoop::from_registry().connected());
     assert!(rita_exit::rita_loop::RitaLoop::from_registry().connected());
 
     system.run();
